+++ packages/react-components/src/lib/components/index.ts
@@ -1,6 +1,8 @@
 export * from './containers';
+export * from './media';
 export * from './page-headers';
-export * from './page-heroes';
+export * from './page-sections';
 export * from './page-widgets';
 export * from './typography';
 export * from './ui';
+