+++ packages/react-components/src/lib/components/page-sections/index.ts
@@ -1,2 +1,3 @@
 export * from './basic-hero';
 export * from './background-image-hero';
+export * from './youtube-with-content';